Buildfix: one stream_utils
Because we are't using modules (__init__), we can only have one module
of a given name. The duplicate stream_utils caused the sim test and
not_tcp_test to fail.
